cv.tables.pesaran(GVAR)	R Documentation

Estimates of critical values of the limiting distributions of the cointegrating rank statistics in a VECM with weakly exogenous I(1) variables

Description

These data sets provide estimated critical values for the limiting distributions of the cointegrating rank trace (CV.trace.table) and maximum eigenvalue (CV.maxeigen.table) statistics for all five intercept/trend scenarios (case I to V) as describe in PSS 2000.

Usage

data(cv.tables.pesaran)

Format

The format is: Lists of 10 $ case I 5 $ case I 10 $ case II 5 $ case II 10 $ case III 5 $ case III 10 $ case IV 5 $ case IV 10 $ case V 5 $ case V 10

Details

details to follow

Source

M. Hashem Pesaran, Yongcheol Shin, and Richard J. Smith. Structural analysis of vector error correction models with exogenous I(1) variables. Journal of Econometrics, 97:293-343, 2000.

James G. MacKinnon, Alfred A. Haug and Leo Michelis. Numerical Distribution Functions of Likelihood Ratio Tests for Cointegration. Journal of Applied Econometrics, 5(14):563-577, 1999.

Examples

data(cv.tables.pesaran)

CV.trace.table[[1]][12,1]         ## returns the 0.05 trace CV for case I in a ... model
CV.maxeigen.table[[4]][12,1]      ## returns the 0.1 maximum eigenvalue CV for case II in a ... model

[Package GVAR version 0.1 Index]