GVAR(GVAR)	R Documentation

Global Vector Auto-Regressive Modelling

Description

GVAR computes VECMs for all regions and stacks the models to a Global Vector Autoregressive Model

Usage


GVAR(data, tw = NULL, p, q = p, r = NULL, weight, case, exo.var = FALSE,
     d = NULL, endo = NULL, ord = NULL, we = NULL, method = "max.eigen")

Arguments

data 	 timeseries data as list (each entry is a matrix of a subsystem of variables, if exo.var=TRUE the last entry are exogeneous variables)

tw 	 time window, vector of start and end point, if NULL the maximum time interval will be used

p 	 scalar/vector of endogenous lags, if a scalar is provided the same lag length is used for all subsystems

q 	 scalar/vector of (weakly) exogeneous lags, if a scalar is provided the same lag length is used for all subsystems

r 	 scalar/vector of cointegrating relations, if a scalar is provided the same cointegration rank is used for all subsystems, if set to NULL the cointegration rank will be determind by method

weight 	 weight matrix, the diagonal elements need to be zero

case 	 single value/vector of cases ("I" to "V"), where case "I" is a zero intercept, zero trend model, case "II" is a restricted intercept, zero trend model, "III" is a unrestricted intercept, zero trend model, "IV" is a unrestricted intercept restricted trend model and "V" is a unrestricted intercept, unrestricted trend model; if a single value is provided the same structure is used for all subsystems

exo.var 	 if TRUE strictly exogeneous variables are included in the model

d 	 list showing which strictly exogeneous variables enter the subsystem equations, if NULL all variables in the last entry of data will be used for all subsystems

endo 	 list of endogenous variables used in each subsystem, if NULL all variables in data will be used

ord 	 vector used if variables in the different subsystem don't appear in the same order, order of each subsystem is concatenated to one vector, if NULL the variables in data are assumed to be ordered identically in all subsystems

we 	 list with numbers of weakly exogeneous variables included in each VECM, corresponds to numbers in ord, if NULL every variable appearing in all subsystems will be used

method 	 select cointegrating rank by max. eigenvalue ("max.eigen") or trace statistic ("trace")

Details

The function computes a VECM for every subsystem before stacking the results to a GVAR model.

Specification of input here.

Value

An object of class "GVAR" containing the following items:

subsys 	 subsystem names

G 	 multiplier matrix for the current variables

H 	 multiplier matrix for the lagged variables

Upsilon.0 	 multiplier matrix for the current strictly exogenous variables

Upsilon 	 multiplier matrix for the lagged strictly exogenous variables

c.0 	 multiplier matrix for the intercept

c.1 	 multiplier matrix for the trend

we.vecms 	 VECMs of the subsystems

m 	 total variables in the subsystems

n 	 endogenous variables in the subsystems

p 	 endogenous lags in the subsystems

q 	 (weakly) exogenous lags in the subsystems

r 	 cointegration rank in the subsystems

weight 	 weight matrix used to calculate the weakly exogenous variables

U 	 residuals of the GVAR

U.cov 	 residual covariance matrix

Author(s)

Martin Summer, Klaus Rheinberger, Rainer Puhr

References

Stephane Dees, Filippo di Mauro, Hashem Pesaran, and L. Vanessa Smith. Exploring the international linkages of the Euro area: A global VAR analysis. Journal of applied Econometrics, 22(1), 2007.

Soeren Johansen. Likelihood-Based Inference in Cointegrated Vector Auto-Regressive Models. Advanced Texts in Econometrics. Oxford University Press, 1995.

M. Hashem Pesaran, Yongcheol Shin, and Richard J. Smith. Structural analysis of vector error correction models with exogenous I(1) variables. Journal of Econometrics, 97:293-343, 2000.

See Also

est.we.mdls

[Package GVAR version 0.1 Index]