RDp(GVAR)	R Documentation

German long term interest and inflation rate

Description

This data set gives the quarterly German nominal long term interest rate (source: Monatsberichte der Deutschen Bundesbank, quarterly values are values of last month of quarter) and the quarterly German inflation rate (source: Deutsches Institut für Wirtschaftsforschung, Volkswirtschaftliche Gesamtrechnung). The sample period is the second quarter of 1972 to the fourth quarter of 1998.

Usage

data(RDp)

Format

A time series matrix containing 2 columns of 107 observations each.

Details

The data before 1990Q3 is West Germany data only, all of Germany afterwards.

Source

Helmut Luetkepohl. New Introduction to Multiple Time Series Analysis. Springer, 2005.

Examples


data(RDp)
plot(RDp)

[Package GVAR version 0.1 Index]