cv.tables.johansen(GVAR)	R Documentation

Estimates of critical values of the limiting distributions of the cointegrating rank statistics in a VECM

Description

This data set provides estimated critical values for the limiting distributions of the cointegrating rank trace (CV.trace.table) and maximum eigenvalue (CV.maxeigen.table) statistics for all five intercept/trend scenarios (case I to V) as describe in Joh 1995.

Usage

data(cv.tables.johansen)

Format

List of 5 tables. Each table gives the 90%-, 95%- and 99%-quantiles for the cointegration rank test for p variables minus rank r.

Details

table15.1 gives the tables for the zero intercept, zero trend model, table15.2 for the restricted intercept, zero trend model, table15.3 for the unrestricted intercept, zero trend model, table15.4 for the unrestricted intercept, restricted trend model, and table15.5 for the unrestricted intercept unrestricted trend model.

Source

Soeren Johansen. Likelihood-Based Inference in Cointegrated Vector Auto-Regressive Models. Advanced Texts in Econometrics. Oxford University Press, 1995.

Examples

data(cv.tables.johansen)
cv.tables.johansen$table15.1

[Package GVAR version 0.1 Index]