est.vecm.mdls(GVAR)	R Documentation

Estimation of Vector Error Correction Models

Description

est.vecm.mdls will calculate VECMs and provide trace and eigenvalue statistics to determine the cointegration rank of the system

Usage


est.vecm.mdls(Y.ts, etw, p, case, r, season = NULL, season.start.time = NULL)

Arguments

Y.ts 	 a time series data matrix containing the endogenous data

etw 	 estimation time window, a list of the form list(start=,end=,freq=)

p 	 lag order of the endogenous variables

case 	 intercept and trend options from "I" to "V", where case "I" is a zero intercept, zero trend model, case "II" is a restricted intercept, zero trend model, "III" is a unrestricted intercept, zero trend model, "IV" is a unrestricted intercept restricted trend model and "V" is a unrestricted intercept, unrestricted trend model

r 	 cointegrating rank used for estimating the model

season 	 frequency of optional seasonal dummies, default value is NULL (no seasonal dummies)

season.start.time 	 optional dummy start, default value is NULL

Details

Calculating VECMs:

Value

An object of class "vecm" containing the following items

type 	 is set to "VECM" by default

dat 	 data used for calculating the model

freq 	 frequency of the time series used

n 	 number of variables entering the model

p 	 lag order of the endogenous variables

r 	 cointegrating rank used for the model estimation

T 	 length of the time series, given without initial values

alpha 	 list of model coefficients for the different cointegration ranks, see details

beta 	 list of model coefficients for the different cointegration ranks, see details

Pi 	 list of model coefficients for the different cointegration ranks, see details

Gamma 	 list of model coefficients for the different cointegration ranks, see details

case 	 intercept and trend specification from "I" to "V", see arguments

mu.0 	 list of intercepts for the different cointegration ranks, see details

mu.1 	 list of trend coefficients for the different cointegration ranks, see details

Phi 	 list of model coefficients for the different cointegration ranks, see details

Omega 	 list of covariance matrices of the error terms for the different cointegration ranks

residuals 	 model residuals

S 	 product moment matrices

lambda 	 eigenvalues

se 	 standard errors of coefficients

tvalues 	 t-distributed test statistic of coefficients

pvalues 	 p-values for test statistic of coefficients

Note

~~further notes~~

Author(s)

Martin Summer, Klaus Rheinberger, Rainer Puhr, Michael Sigmund

References

Soeren Johansen. Likelihood-Based Inference in Cointegrated Vector Auto-Regressive Models. Advanced Texts in Econometrics. Oxford University Press, 1995.

Helmut Luetkepohl, Markus Kraetzig. Applied Time Series Econometrics. Cambridge University Press, 2004.

Helmut Luetkepohl. New Introduction to Multiple Time Series Analysis. Springer, 2005.

See Also

est.we.mdls

Examples


 data(RDp)

    dt <- 1/tsp(RDp)[3]
     k <- 4
   etw <- list(start=tsp(RDp)[1]+k*dt,end=tsp(RDp)[2],freq=tsp(RDp)[3])

 model <- est.vecm.mdls(Y.ts=RDp,etw=etw,p=k,r=1,case="III",season=4,season.start.time=1)
 summary(model)

[Package GVAR version 0.1 Index]