est.we.mdls(GVAR)	R Documentation

Estimation of Weakly Exogenous VEC Models

Description

est.we.mdls will calculate weakly exogneous VECMs and provide trace and eigenvalue statistics to determine the cointegration rank of the system

Usage

est.we.mdls(z.ts, etw, p, q = p, n, case, r)

Arguments

z.ts 	 a time series data matrix containing endogenous and weakly exogenous variables

etw 	 estimation time window, a list of the form list(start=,end=,freq=)

p 	 lag order of the endogenous variables

q 	 lag order of the (weakly) exogenous variables

n 	 number of endogenous variables

case 	 intercept and trend options from "I" to "V", where case "I" is a zero intercept, zero trend model, case "II" is a restricted intercept, zero trend model, "III" is a unrestricted intercept, zero trend model, "IV" is a unrestricted intercept restricted trend model and "V" is a unrestricted intercept, unrestricted trend model

r 	 cointegrating rank used for model estimation

Details

Calculating weakly exogneous VECMs:

Value

An object of class "vecm" containing the following items

type 	 is set to "weakly exogenous VECM" by default

dat 	 data used for calculating the model

freq 	 frequency of the time series used

m 	 number of total variables entering the model

n 	 number of endogenous variables entering the model

p 	 lags for the endogenous variables

q 	 lags for the (weakly) exogenous variables

r 	 cointegrating rank used for model estimation

T 	 length of the time series, given without initial values

alpha 	 list of model coefficients for the different cointegration ranks, see details

beta 	 list of model coefficients for the different cointegration ranks, see details

Pi.y 	 list of model coefficients for the different cointegration ranks, see details

Phi 	 list of model coefficients for the different cointegration ranks, see details

Psi 	 list of model coefficients for the different cointegration ranks, see details

Lambda 	 list of model coefficients for the different cointegration ranks, see details

case 	 intercept and trend specification from "I" to "V", see arguments

c.0 	 list of intercepts for the different cointegration ranks, see details

c.1 	 list of trend coefficients for the different cointegration ranks, see details

Omega.uu 	 list of covariance matrices of the error terms for the different cointegration ranks

S 	 product moment matrices

lambda 	 eigenvalues

residuals 	 model residuals

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

Martin Summer, Klaus Rheinberger, Rainer Puhr

References

Soeren Johansen. Likelihood-Based Inference in Cointegrated Vector Auto-Regressive Models. Advanced Texts in Econometrics. Oxford University Press, 1995.

M. Hashem Pesaran, Yongcheol Shin, and Richard J. Smith. Structural analysis of vector error correction models with exogenous I(1) variables. Journal of Econometrics, 97:293-343, 2000.

See Also

est.vecm.mdls

Examples

##---- Should be DIRECTLY executable !! ----

[Package GVAR version 0.1 Index]