rank.test.vecm(GVAR)	R Documentation

Test for Cointegrating Rank in Vector Error Correction Models

Description

rank.test.vecm performs a test for cointegrating rank in VEC models

Usage


rank.test.vecm(Y.ts, etw, p, case, season = NULL, season.start.time = NULL)

rank.test.we(z.ts, etw, p, q = p, n, case)

Arguments

Y.ts,z.ts 	 time series data matrix

etw 	 estimation time window, a list of the form list(start=,end=,freq=)

p 	 lag order of the endogenous variables

q 	 lag order of the (weakly) exogenous variables, is set to q=p by default

n 	 number of endogenous variables

case 	 intercept and trend options from "I" to "V", where case "I" is a zero intercept, zero trend model, case "II" is a restricted intercept, zero trend model, "III" is a unrestricted intercept, zero trend model, "IV" is a unrestricted intercept restricted trend model and "V" is a unrestricted intercept, unrestricted trend model

season 	 optional seasonal dummies, default value is NULL

season.start.time 	 optional dummy start time, default value is NULL

Details

Either use rank.test.vecm or rank.test.we for VECM and VECM with exogenous I(1) variables respectively.

Value

Values for maximum eigenvalue and trace statistics are returned as well as critical values for the limit distribution of these test statistics.

Note

~~further notes~~

Author(s)

Martin Summer, Klaus Rheinberger, Rainer Puhr

References

Soeren Johansen. Likelihood-Based Inference in Cointegrated Vector Auto-Regressive Models. Advanced Texts in Econometrics. Oxford University Press, 1995.

Helmut Luetkepohl, Markus Kraetzig. Applied Time Series Econometrics. Cambridge University Press, 2004.

M. Hashem Pesaran, Yongcheol Shin, and Richard J. Smith. Structural analysis of vector error correction models with exogenous I(1) variables. Journal of Econometrics, 97:293-343, 2000.

See Also

est.vecm.mdls, est.we.mdls

Examples


##---- Should be DIRECTLY executable !! ----

[Package GVAR version 0.1 Index]