set.mdl(GVAR)	R Documentation

Transform VECM to VAR

Description

Transform to the VAR view from a VECM model calculated with est.vecm.mdls or est.we.mdls

Usage


set.mdl(mdls, exo = NULL, skip = NULL)

Arguments

mdls 	 an object of class "vecm"

exo 	 number of exogenous variables included in the I(1)-model, default is NULL

skip 	 skip model parameters that should not be converted to VAR view, default is NULL

Value

A list containing the VECM and VAR views of the model calculated

Author(s)

Martin Summer, Klaus Rheinberger, Rainer Puhr

References

Stephane Dees, Filippo di Mauro, Hashem Pesaran, and L. Vanessa Smith. Exploring the international linkages of the Euro area: A global VAR analysis. Journal of applied Econometrics, 22(1), 2007.

Soeren Johansen. Likelihood-Based Inference in Cointegrated Vector Auto-Regressive Models. Advanced Texts in Econometrics. Oxford University Press, 1995.

M. Hashem Pesaran, Yongcheol Shin, and Richard J. Smith. Structural analysis of vector error correction models with exogenous I(1) variables. Journal of Econometrics, 97:293-343, 2000.

Examples


##---- Should be DIRECTLY executable !! ----

[Package GVAR version 0.1 Index]