ur.tests(GVAR)	R Documentation

Unit Root Tests

Description

ur.tests computes the three different unit root test statistics, Augmented Dickey-Fuller Test, Phillips-Perron Unit Root Test and Kwiatkowski-Phillips-Schmidt-Shin (KPSS) Test, from tseries for a data matrix.

Usage


ur.tests(X)

Arguments

X 	 a (time series) data matrix

Value

Returns a matrix giving the alternative hypothesis of the tests and the p-values for each variable.

Author(s)

Martin Summer, Klaus Rheinberger, Rainer Puhr

References

see package tseries

Examples


X <- ts(cbind(rnorm(100),rnorm(100),rnorm(100)),start=1980,freq=4)
ur.tests(X)

data(RDp)
ur.tests(RDp)

[Package GVAR version 0.1 Index]