we.diag(GVAR)	R Documentation

Exogeneity diagnostics

Description

This function tests variables for weak exogeneity.

Usage

we.diag(Z, n, p, q = p, case)

Arguments

Z 	 a time series data matrix

n 	 number of endogenous variables, the remaining variables will be tested for weak exogeneity

p 	 lag order of the endogenous variables

q 	 lag order of the weakly exogenous variables

case 	 intercept and trend options from "I" to "V", where case "I" is a zero intercept, zero trend model, case "II" is a restricted intercept, zero trend model, "III" is a unrestricted intercept, zero trend model, "IV" is a unrestricted intercept restricted trend model and "V" is a unrestricted intercept, unrestricted trend model

Details

In order to be weakly exogenous variables X must be (a) integrated of order 1 (X~I(1)); it follows (b) that X is not cointegrated on its own, and (c) that the differenced process does not depend on the lagged Z.

The variables tested will be the variables from the last (N-n) columns of the matrix Z, where N is the total number of columns.

Value

Results for test (a) to (c).

Author(s)

Martin Summer, Klaus Rheinberger, Rainer Puhr

References

Stefan Zeugner. Implementing Pesaran-Shin-Smith. First year paper, Institute for Advanced Studies, Vienna, 2006.

Soeren Johansen. Likelihood-Based Inference in Cointegrated Vector Auto-Regressive Models. Advanced Texts in Econometrics. Oxford University Press, 1995.

M. Hashem Pesaran, Yongcheol Shin, and Richard J. Smith. Structural analysis of vector error correction models with exogenous I(1) variables. Journal of Econometrics, 97:293-343, 2000.

See Also

rank.test.vecm

Examples

##---- Should be DIRECTLY executable !! ----

[Package GVAR version 0.1 Index]